This file to get the data from ACAPS API
Its usage is quite simple and everything is mentioned in the comments. When you need to use any data set just change the `endpoint` and everything should work. I also had some checking blocks to make sure if the token was generated correctly or if something went wrong and will help figure out where the problem is. Don't forget to create an account on ACAPS to have the username and password also note that the username is the email you will register with.
Note: It will save the data on a data frame and you have a line to save the data as a CSV file too you can comment on this line if you don't want to download the data